feat: made mode upload into a dialog instead its own page
